code/coredms/test/requirements.txt
pika==1.3.1
cassandra-driver
